pyproject.toml
[project]
name = "projects"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "arxiv>=2.2.0",
    "google-api-python-client==2.124.0",
    "google-auth-oauthlib==1.2.0",
    "google-auth==2.29.0",
    "isodate>=0.7.2",
    "langchain>=0.3.26",
    "langchain-community>=0.3.27",
    "langchain-core>=0.3.69",
    "langchain-groq>=0.3.6",
    "langgraph>=0.5.3",
    "langgraph-cli[inmem]>=0.3.5",
    "playwright>=1.54.0",
    "pydantic>=2.11.7",
    "python-dotenv>=1.1.1",
    "streamlit>=1.47.0",
    "tweepy>=4.16.0",
    "wikipedia>=1.4.0",
    "tiktoken>=0.9.0",
    "httpx==0.27.0",
]
